Yathzee REXX v1.00
The next idea was to create a game myself, and this is when it started
to be more fun.

We regularly played a yathzee game at home with some dices and score
papers where you needed to fill in what was going on. I decided to
create the same game on the computer.

My father was 30 years older than me, and started working in 1975 as a
cobol programmer and system programmer (mainframe stuff I guess) at a
dutch insurance company. This was around the time that the IBM mainframe
got the first remote terminal connections to another buildings, using
phone lines, coax lines and binary protocols over 9600 bps connections
(IBM Systems Network Architecture), so they didn't have to drive around
all day any more with cars carrying buckets of punch cards to do
insurance data lookups and mutations. :D

20 years later, when I started programming, that insurance company was
already merged with another bigger insurance company with an orange logo
and a bank with a blue logo. The result had an orange logo, and my dad
was managing software releases to production systems, keeping them
running, doing on-call shifts and failover tests of their production
systems to another location in the country etc...

So, long story short, apparently one of the first things he taught me
was that you need to version your code and have a release process to
take it into production. That's why started labeling this first piece of
code with a v1.00 label. :)

Two versions later the changelog would appear, with dates when my game
code would have been taken "into production". Hahaha.

Fun thing is, I'm doing exactly the same kind of work now, be it at a
very different company, and again 20 years later. :-]

After writing this first version of the Yathzee game, I immediately
realized that I had to learn new things, to get rid of the ugly
repetitive code, reaized that REXX didn't handle doing screen input and
output for this kind of applications very well (e.g. you can just keep
typing input even after you cross the end of the input field).

But hey! I implemented bubblesort! :-D

And I wrote this in 1995, but the last modification time of the file is
in 1996.