YATHZEE REXX Versie 2.11
The score top 10 became a Top 11, probably because a Top 10 was too
boring.

I was still searching for a better way to do user input. This version
replaces Pull with Linein, because with the Pull command, the up and
down keys would scroll through the command history of DOSKEY.

Also, attached two screenshots again.