YATHZEE Versie 2.12
Most of the diff is cosmetics, changing the case of letters in
subroutine / function calls.

Ah, and by providing a command line argument, it's possible to use
another Top 11 score list: scr=Top11File(Arg(1))

And I added an ascii art picture of the program flow. I think it was
helpful, because on the 80x25 text screen, you could easily get lost in
a file of almost 800 lines...