REXXPSP - Exploring the Program Segment Prefix
When calling external functions from REXX, the first hundred bytes in
memory are filled with information. Since the .COM (or renamed to .RX)
program starts at offset 100h, we can look in the memory between 0 and
FFh to see what we got.

This program dumps the first 100h bytes of memory to a file,
REXXPSP.HEX, so it can be inspected.