Extended RXCRSTAT: reverse engineering!
REXX has an external function RXCRSTAT which can be used to show or hide
the blinking cursor. But, it can't tell you the current state.

To be able to do this, I disassembled the RXCRSTAT.RX from PC DOS (see
RXCRSTAT.TXT) and made an extended version to replace it.