USRINPUT.ASM: v2.00, user input like it was meant to be!
"Tired of (Parse) Pull, RxGetKey and CharIn ?  Try this!"

This is the successor to USRINPUT.BAT, but written in assembler.

This input helper (again) has the following features:
- specify amount of characters to input
- text to pre-fill the input field with
- show or don't show the cursor
- echo or don't echo the typed characters
- force all input to uppercase
- insert or replace characters (?)
- does the user need to press enter or not?
- delete pre-filled default text on first edit?
- what character to display for the empty part of the input field
- a character to display instead of the input (e.g. to build a password
  field with ***)

Try PROB.BAT and you'll see an edit field, with bold green text on a
blue background, dots for empty places in the input field, a blinking
cursor, etc...

UsrInput('301B1A2E'x,'hallo')

hallo...........................................

See USRINPUT.TXT for a manual (in dutch).
Also see extra/prob.png for a screenshot.